+++ BiliLiveVisual/Assets/Scripts/3rd/BiliLive/BiliListener.cs
@@ -62,7 +62,7 @@ public class BiliLiveListener
                 var info = jsonData["info"];
                 outData = new BiliLiveDanmakuData.DanmuMsg
                 {
-                    uid = int.Parse(info[2][0].ToString()),
+                    uid = long.Parse(info[2][0].ToString()),
                     nick = info[2][1].ToString(),
                     color = "#" + int.Parse(info[0][3].ToString()).ToString("x6"),    //十进制转十六进制字符串
                     content = info[1].ToString()
@@ -73,7 +73,7 @@ public class BiliLiveListener
                 var data = jsonData["data"];
                 outData = new BiliLiveDanmakuData.SendGift
                 {
-                    uid = int.Parse(data["uid"].ToString()),
+                    uid = long.Parse(data["uid"].ToString()),
                     uname = data["uname"].ToString(),
                     action = data["action"].ToString(),
                     giftName = data["giftName"].ToString(),
@@ -84,7 +84,7 @@ public class BiliLiveListener
                 var data = jsonData["data"];
                 outData = new BiliLiveDanmakuData.ComboSend
                 {
-                    uid = int.Parse(data["uid"].ToString()),
+                    uid = long.Parse(data["uid"].ToString()),
                     uname = data["uname"].ToString(),
                     action = data["action"].ToString(),
                     combo_num = int.Parse(data["combo_num"].ToString()),
@@ -99,7 +99,7 @@ public class BiliLiveListener
                 var data = jsonData["data"];
                 outData = new BiliLiveDanmakuData.GuardBuy
                 {
-                    uid = int.Parse(data["uid"].ToString()),
+                    uid = long.Parse(data["uid"].ToString()),
                     username = data["username"].ToString(),
                     guard_level = int.Parse(data["guard_level"].ToString()),
                     price = int.Parse(data["price"].ToString()),
@@ -116,7 +116,7 @@ public class BiliLiveListener
 
                 outData = new BiliLiveDanmakuData.SuperChatMessage
                 {
-                    uid = int.Parse(data["uid"].ToString()),
+                    uid = long.Parse(data["uid"].ToString()),
                     uname = user_info["uname"].ToString(),
                     face = user_info["face"].ToString(),
                     face_frame = user_info["face_frame"].ToString(),
